He approached Andy which startled him and make him pee everywhere on the floor and on Woody too.
At this point in the story, Woody has given up the charade of being just a toy and has confronted Andy directly. Discovering that one's inanimate toys were in fact, capable of independent movement would be shocking to anyone, and Andy is no exception. Indeed, Andy is so startled that he loses control of his bladder. Cynics frequently point out that human males cannot urinate when erect, as the external urethral sphincter closes, however this is not always true as the internal urethral sphincter (which is closer to the bladder) remains open until orgasm, thus it is possible, though unlikely, to spontaneously urinate if the external urethral sphincter receives a sudden shock. In this case, so much urine was accidentally released it flowed "everywhere on the floor" as well as on Woody himself. The urine of course did not literally go "everywhere," that is merely an overstatement for dramatic effect.