Being drenched in his urine made him harder than ever!
This sentence has confused many readers, as it is unclear which of the characters is becoming more "hard" or having a more powerful erection. The issue is primarily one of grammar, but the debate is very important in the events of the story. There are two camps among scholars, each believing it refers to difference characters. The A-Urolagniaists believe that the sentence is to be read as a stand alone and Andy is the subject of the both pronoun. In this interpretation, Andy has become more aroused after being drenched in his own urine. W-Urolagniaists believe that the oblique pronoun "him" refers to Woody, as Woody was the last subject directly mention. In this interpretation, after being covered in Andy's urine, Woody became more aroused. W-Urolagniaists argue that since Woody is the only character directly mentioned to have urine on him, he must be the one getting harder. A-Urolagniaists counter by saying that the author has omitted specifics several times in the work, and it implied that while peeing with an erection in a sitting position, Andy would have gotten some of his urine on himself. The debate rages in Space Universities even today.