I present to you all, a fully customized ghost role known as the Space Police!
The Space Police, was an idea inspired by a comment made from Dex. Why not have a Ghost Role that still makes you follow alongside Nanotrasen? So I decided to get too work. Remembering an unused and very dope looking Police outfit existed. I went ahead and coded in my own ghost role spawner thus introducing the "Space Police", included with a very very heavily nerfed space bike called "The Law". Now when chasing an antag down you can say- "The Law is gaining on you..."

But the Space Polices goal is to in short, hunt down antagonists, whether it be on the station with the watchful eye of the crew and those in command- (As those in Command have full jury over this Policeman) -or out deep in space! Course whether the Policeman decides to take a confirmed Antagonist dead or alive is there choice. Though under any circumstances are they aloud to harm the crew. Period. In fact if Command so chooses they can ask for the Policeman to be exterminated and the Policeman must sit back and let what's done be done. They come equipped with a basic energy rifle and energy pistol. With there only benefit being they can live in space! Still need oxygen but you can ride The Law in style! Also this can give a hard counter to Xeno-Lings, making being a Xeno-Ling actually dangerous. Along with being a Syndicate Comms Agent.