Fixed the Space Police map and added a new Ghost Role based along Xenobiology!
So, I was talking with a friend and they thought- "Wouldn't it be cool to have a ghost role that specifically revolves around Xenobiology? But you wanna take revenge on Nanotrasen?" -and I thought that sounded super dope! So with him shooting out ideas I got to work on a full Xenobiology related Ghost Role. Everything from making power to getting yourself food will require you to do Xenobiology. It will be a fun challenge for people who love Xeno and want a- Tougher time with an end goal... Spread your slimes EVERYWHERE.