SpaceG Fix (Again)
Forgot to remove the ladder. Which why it was removed is I just tested the second floor a few moments ago only to see the second floor was placed in Lavaland... Somehow and that's a big NO. So no-second floor. I will be making just the second floor it's own map soon.